swap to interfaces and package private impls
