+++ api/src/main/java/team/unnamed/creative/metadata/gui/GuiBorder.java
@@ -24,74 +24,16 @@
 package team.unnamed.creative.metadata.gui;
 
 import net.kyori.examination.Examinable;
-import net.kyori.examination.ExaminableProperty;
-import net.kyori.examination.string.StringExaminer;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.Objects;
-import java.util.stream.Stream;
+public interface GuiBorder extends Examinable {
 
-public class GuiBorder implements Examinable {
+    int top();
+    int bottom();
+    int left();
+    int right();
 
-        private final int top;
-        private final int bottom;
-        private final int left;
-        private final int right;
-
-        private GuiBorder(int top, int bottom, int left, int right) {
-            this.top = top;
-            this.bottom = bottom;
-            this.left = left;
-            this.right = right;
-        }
-
-        public int top() {
-            return top;
-        }
-
-        public int bottom() {
-            return bottom;
-        }
-
-        public int left() {
-            return left;
-        }
-
-        public int right() {
-            return right;
-        }
-
-    @Override
-    public @NotNull Stream<? extends ExaminableProperty> examinableProperties() {
-        return Stream.of(
-                ExaminableProperty.of("top", top),
-                ExaminableProperty.of("bottom", bottom),
-                ExaminableProperty.of("left", left),
-                ExaminableProperty.of("right", right)
-        );
+    static @NotNull GuiBorder of(int top, int bottom, int left, int right) {
+        return new GuiBorderImpl(top, bottom, left, right);
     }
-
-    @Override
-    public String toString() {
-        return examine(StringExaminer.simpleEscaping());
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof GuiBorder)) return false;
-        GuiBorder guiBorder = (GuiBorder) o;
-        return top == guiBorder.top && bottom == guiBorder.bottom && left == guiBorder.left && right == guiBorder.right;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(top, bottom, left, right);
-    }
-
-    public static GuiBorder of(int top, int bottom, int left, int right) {
-        return new GuiBorder(top, bottom, left, right);
-    }
-
-
 }
